Simple device plugin 'abstract' implementation.
The NSM is going to need to provide a 'deviceplugin' for
a resource

nsm.ligato.io

so that Pods can request NSM as a resource.

The NSM is also going to provide a 'deviceplugin' for
resources:

nsm.ligato.io/${service}/${channel}

so that the Pod may request an actual hardware NIC providing
that Network Service.

In otherwords, the NSM will need to provide multiple 'deviceplugins'.

For this reason, I'm attemtping to abstract he housekeeping parts
of the 'deviceplugin' into a deviceplugin.go and we can then use
Go composition to construct easily the *specific* device plugins we
need for those resources.

References:
"Why Go Structs are superior to class based inheritance":
https://medium.com/@simplyianm/why-gos-structs-are-superior-to-class-based-inheritance-b661ba897c67

K8s DevicePlugin documentation:
https://kubernetes.io/docs/concepts/cluster-administration/device-plugins/

Signed-off-by: Ed Warnicke <hagbard@gmail.com>
Signed-off-by: Ed Warnicke <eaw@cisco.com>

Signed-off-by: Stanislav Chlebec <stanislav.chlebec@pantheon.tech>