Very rough untested first pass at the NSM Device Plugin
This code builds, but has not been tested.  Basically,
it creates a mountdir to be injected into the Pod requesting
an NSM device.  There is a FIXME comment indicating where
we should be creating the socket file in that subdir for
the Pod->NSM communication.

There's also a fair bit of handwaving here, because the Pod
mounts come from teh Host.  Since the NSM is expected to be run
in a container, we need to make sure that
/var/lib/networkservicemesh is mounted into the NSM, so that
we can mount
/var/lib/networkservicemesh/${id}/ into the Pod that requested
an NSM.

Signed-off-by: Ed Warnicke <eaw@cisco.com>