Convert networkservicemesh into a Ligato plugin
This commit moves netmesh over the ligato framework to take advantage
of the many things the framework provides. Specifically, we take
advantage of being able to run the netmesh code as a Ligato plugin,
allowing the framework to setup our control loop.

Additionally, this commit changes the name of the docker image we
build to be ligato/networkservicemesh/netmesh, removing the older
ligato/networkservicemesh/nsm image name.

This commit includes instructions for running netmesh as a daemonset
as well.

Co-Authored-By: John Burns <johnburn@cisco.com>