Utilities to make idempotent plugins easy
Embedding

idempotent.Impl

In a plugin and writing the Plugin Init() and Close()
as:

func (plugin *Plugin) Init() error {
     return plugin.IdempotentInit(plugin.init)
}

func (plugin *Plugin) Close() error {
     return plugin.IdempotentClose(plugin.close)
}

will make sure that the plugin Init() and Close()
are idempotent.  Additionally, refCounting is performed
so that the Plugin isn't *actually* closed until
Close() has been called as many times as Init()

This allows safe sharing of a Plugin.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>