First swag at wrapping grpc.Plugin
A few notes:
First, this isn't done.

Second, I have hacked a file in vendor
until I can get a patch upstream.  Be not alarmed,
this shouldn't be merged till that's resolved.

What this patch does primarily is to wrap an
idempotent Plugin around grpc.Plugin from cn-infra

It also adds the NewPlugin() SharedPlugin()
pattern in options.go.

I'm currently thinking of ways to make SharedPlugin()
a multi-singleton, so that if you call

SharedPlugin(options...) you get a Plugin that has
the deps and config provides by those options, and
if someone has asked for one like that before you
you get the Shared one.  But for now, I only have
the Singleton from the one that is using Default
Deps and Config.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>