Idempotent logger and config plugins
This commit adds idempotent logger and config plugins.

The config is based on Viper, and makes use of cobra

The logger is based on logrus

logger has 100% code coverage

config is *almost* 100%, but has an err check for a place
where an err is returned, but the underlying source can't
possibly return a non nil err.  But since it could in the
future, the err is checked, but literally can't be tested.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>