Update minikube version to v0.28.0 (#133)
* Update minikube version to v0.28.0 and kubernetes to v1.11.0

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Leave kubernetes at v1.10.0

The Kubernetes v1.11.0 bump will happen in a separate commit.

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Use the localkube bootstrapper

It appears that versions of minikube >= v0.26 require systemd. Travis
does not have systemd in it's images, thus minikube fails. A workaround,
as documented by [1], is to pass "--bootstrapper=localkube" on the
minikube commandline. But localkube will go away in a future Kubernetes
release, so this is only a workaround.

[1] https://github.com/kubernetes/minikube/issues/2704

Signed-off-by: Kyle Mestery <mestery@mestery.com>