Add proxy support to docker build commands in the Makefile
When trying to use the Makefile to build on a machine behind a proxy, it
was observed the docker build commands all failed. This commit adds support
for using a proxy with the docker build commands in the Makefile. It also
fixes an explicit usage of "wget" in the nsm Test.Dockerfile to use a proxy.

Signed-off-by: Kyle Mestery <mestery@mestery.com>