Add CheckDeps, InitDeps, CloseDeps helper (#170)
* Add CheckDeps, InitDeps, CloseDeps helpers

CheckDeps - check for common errors in plugin Deps

The Deps pattern in Plugins is a powerful way to allow
users to wire in alternate implementations for things the Plugin
uses.  In addition to the flexibility it offers, it also makes
it much much easier to Mock and test.

Unfortunately, there are several mistakes its easy to make with Deps:

- Making a Deps field private (lowercase) which precludes wiring in
  an alternate implementation
- Failing to set a Default value for Deps in options.go

CheckDeps checks for all of these mistakes.  It is suitable to
be run in the NewPlugin(...) in options.go as a final check before
returning a new Plugin to insure its ready to Go :)

*******

InitDeps - Runs Init() on all Deps implemented with Plugins

Deps should be interfaces, they do not have to be Plugins
InitDeps loops through all of the fields in Deps, and if they
are Plugins, it Init()'s them.

CloseDeps - Runs Close() on all Deps implemented with Plugins

Closes all Deps of the given plugin if they themselves
are idempotent.Plugins


* Fixes in response to review comments


* Minor requested fix to comments


* One more comment to resolve