Add the ability to publish images to Docker Hub
Closes #186

This commit adds the ability to publish Docker Hub images into the
`networkservicemesh` Docker Hub organization. The following images are
published with this commit:

- networkservicemesh/nsm-daemon
- networkservicemesh/nsm-simple-dataplane
- networkservicemesh/nsm-init
- networkservicemesh/nse

Note that in addition, this commit revamps the Makefile a bit, moving
some complexity out into a .nsm.mk file which is then included by the
top-level Makefile.

Note also that for now, this commit makes use of my own Docker Hub login
information for build publishing. These are the secrets which are stored
in the .travis.yml file.

Signed-off-by: Kyle Mestery <mestery@mestery.com>