Correct environment variable usage when pushing docker images
In a Makefile, it turns out it's tricky to set environment variables in a
target and then use them. This was being done in .nse.mk by setting $REPO
to whatever repository was being worked on, and it was failing: it was
never set.

It turns out this is an unnecessary optimization, so simply removing this
should allow the `docker push` operations to succeed.

Signed-off-by: Kyle Mestery <mestery@mestery.com>