Rename simple-dataplane to test-dataplane
To better reflect the reality the in-tree dataplane was merely for testing
purposes in our CI system, rename it to `test-dataplane`. This will hopefully
avoid any confusion around whether or not you should run this anywhere near a
production system (you shouldn't), and what it is used for (it's used for
testing only).

Signed-off-by: Kyle Mestery <mestery@mestery.com>