Update Gopkg.lock
According to this blog post [1], Dependabot does support golang with dep,
but it does not quite yet support vendored dependencies. This means it
will update Gopkg.[lock,toml] for us.

In practice, this means when we see a Dependabot PR, someone should pull
it down and manually run `dep ensure` to update our vendored
dependencies.

Note that Dependabot is fixing this behavior and hopefully soon we won't
have to do this anymore.

For now, this updates our vendor directory to match what Dependabot has
done for us.

[1] https://dependabot.com/blog/go-support

Signed-off-by: Kyle Mestery <mestery@mestery.com>