Add yamllint to our travis-ci runs (#263)
* Add yamllint to our travis-ci runs

Closes #262

Since we are making heavy use of yaml files for Network Service Mesh, we should
ensure these are syntactically correct. We'll use yamllint [1] for this.

This commit adds running of yamllint into our travis-ci runs. It also adds a
custom configuration file which turns the yamllint line length check into a
warning. And finally, it fixes a large amount of yamllint errors in our
existing yaml files.

[1] https://github.com/adrienverge/yamllint

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Use pip to install yamllint instead of apt

Since our travis nodes are 14.04 and not 16.04, we have to use pip
to install yamllint instead of apt-get.

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Use our local yamllint configuration file

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Fix lint errors in the configuration file

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Do not run yamllint with --strict

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Fix some yaml files

Signed-off-by: Kyle Mestery <mestery@mestery.com>