WIP DO NOT MERGE Update the CI to use the latest minikube and kubernetes
Closes #311

This updates the CI to use the latest minikube (v0.29.0) and kubernetes
(v1.12.0) for testing.

NOTE: I expect this to fail at the moment since the kubeadm bootstrapper seems to
have some issues here. Pushing this out to see if it fails the same way as it does
in my local environment.

Signed-off-by: Kyle Mestery <mestery@mestery.com>