Add template circleci integration
The plan is to use circle-ci as a test runner to hop over to packet.net
and run tests on real hardware. For now, this puts a template in to get
the hook to circle-ci. We'll want to replace this with the actual steps
to run on packet.net once it merges.

Signed-off-by: Kyle Mestery <mestery@mestery.com>