Copy images from outer to inner Docker
We're using kubeadm-dind-cluster now to run CI tests in travis-ci. This
means we first build the Docker images in the outer Docker, and then
spinup a Kubernetes cluster in the outer Docker, with it's own inner
Docker.

This PR ensures we're copying the images from the outer Docker into the
inner Docker of each Kubernetes node so we run tests against the freshly
built images rather than the ones from Docker Hub.

Signed-off-by: Kyle Mestery <mestery@mestery.com>