Make packet-destroy more stable
packet-destroy as part of the packet-integraion-tests job
was slow and unstable, due to issue with pulling
the provision docker image from the remote_docker
being very very slow (sometimes taking 30 minutes, or
timing out).

This same docker image, when used as the basis for a standalone
job (as is done for packet-deploy) is zippy and reliable.

This patch breaks packet-destroy out into a separate job.
In the event that integration-tests fails, it will use
the CircleCI API to launch the packet-destroy job.

This is done by storing the /cncf/data needed in a cache
with key cncf-data-{{.Environment.CIRCLE_WORKFLOW_ID}}
and having the API set CIRCLE_WORKFLOW_ID in the packet-destroy
job to match.

This has been tested and works.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>