Add a cleanup step to build
This is far from perfect, as it basically waits
10 minutes for packet-deploy to finish before triggering
packet-destroy.  I hate sleeps in tests... but its the best
we can do.

This has been tested:

1.  If the build is broken, cleanup happens correctly.
2.  If the build is broken *and* packet-deploy is broken,
    the packet-destroy triggered by the broken build
    does nothing, and reports success unless packet-deploy
    has completed.

So the only available mode in which this doesn't behave well is:

packet-deploy succeeds more than 10 minutes after the build breaks.
In that case we can leak servers.

The additional drawback is that build fails will take a bit longer to report.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>