Break out log collection to obvious seperate step in CI
Right now, if we break, the 'point of breakage' ends up
being buried behind all the log collection.

The log collection is epic.  This patch breaks log collection out
into a separate script:

./scripts/dump_logs.sh

And puts it into separate steps into TravisCI and CircleCI that *always*
collect logs (error or not).

Signed-off-by: Ed Warnicke <hagbard@gmail.com>