Configure vpp to not consume entire core
VPP by default will 'poll' continuously for traffic.
If you are in a high scalability/performance case this
is important and good.

If you are not it just becomes a process burning 100% of a CPU.

This config sets vpp up to poll-sleep, which radically reduces
CPU usage when not under heavy load, though with some costs
in the heavy load case.

For the average users, this setting is best.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>