Fix vppagent-firewall-nse
So... here's what happened.

When we request a Connection, it comes back to us with the workspace
parameters set.

ConnectionConverter interprets the socketfile fully qualified directory as:

path.Join(baseDir,Connection.GetMechanism().GetParameters()[connection.Workspace],connection.GetMechanism().GetParameters()[connection.Socketfile])

This is correct for the Dataplane, but for an NSE we need it to be:

path.Join(baseDir,Connection.GetMechanism().GetParameters()[connection.Socketfil\
e])

Signed-off-by: Ed Warnicke <hagbard@gmail.com>