Enable the libvirt vagrant provider
The libvirt vagrant provider uses sshfs to sync the folders between the host and the guest.
We add /vagrant as it's needed for the libvirt provider.

Ubuntu 18.04 uses systemd-resolver, which causes issues with dnsmasq and breaks DNS within the guest, which is mandatory for libvirt to operate. Therefore we use 'generic/ubuntu1604' when libvirt is enabled. 

Signed-off-by: Mathieu Rohon <mathieu.rohon@orange.com>
Signed-off-by: Stanislav Chlebec <stanislav.chlebec@pantheon.tech>