Convert to using hostNetworking:true (#791)
* Convert to using hostNetworking:true

This temporarily involves switching to edwarnicke/vpp-agent:v1.8.1 images
while the ligato team spins new updates with fixes in vpp to workaround
a kernel issue.

The basic kernel issue hit was this:

When running in hostNetwork:true, mmap for Hugepages returns without an error
(even though no Hugepages are available to be mmaped) and then the kernel
throws a SIGBUS when we try to use them.  This should *never* happen, but
does.  To work around it, we employ an alternate means to detect that
we don't actually have the pages involving mincore.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>

* Add a comment explaining ACL on mgmt interface.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>

* Update to pull EgressInterface (formerly MgmtInterface) into a seperate struct

Signed-off-by: Ed Warnicke <hagbard@gmail.com>