Adding initial support for kind.sigs.k8s.io
K8s in Docker is a standard way to run K8s tests. We use this tool
here as an alternative cluster provider. Kind is lighter than bringing
up full VMs.

The setup consists of master and 2 nodes. It successfuly runs
make k8s-integration test, so can be used in the CI.

Signed-off-by: Nikolay Nikolaev <nnikolay@vmware.com>