Simplification of SDK
This PR simplifies the SDK, by using the context.Context to pass
data between Endpoints in the chain.

By doing this, and a few other minor changes, the Endpoints used
in CompositeEndpoint just need to be normal implementers of

networkservice.NetworkServiceServer

Among other things, this means that CompositeEndpoints can be nested
within each other.

Note: Updates to sdk/README.md will be forthcoming.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>