Fix yamllint so that we have *no* warnings
Trying to fish actual errors out of the warnings is really
annoying.  This patch fixes the warnings, and shifts to
--strict to fail on warnings.

Also puts the file match for yaml files into the .yamllint.yaml file
so we can just run against.

Also switched to installing with pip to get recent version of
yamllint that supports --strict flag.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>