Transition to using a more conventional FromEnv() approeach to configuration in sdk
Previously, the SDK had scattered calls to:

configuration.CompleteNSConfiguration throughout the various Endpoint pieces.

Convention in many other projects is to using FromEnv().

This patch transitions to using a configuration.FromEnv() to fill in any config from Env *before*
passing the config to the constructor.  This allows the config to be filled out however makes sense to the
NSE author.

The same was done for FromNSUrl

Finally, some endpoints in vppagent/ that were not using common.Configuration at all dropped the argument.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>