Kind uses context now, drop KUBECONFIG (#1938)
* Make kind create cluster wait for the control plane

Use the --wait flag to let kind create cluster to stop
and wait for the clontrol plane to appear. Deafult to
120s which should be sane maximum.

Signed-off-by: Nikolay Nikolaev <nnikolay@vmware.com>

* Kind now usese contexts for kubectl

Remove all references to the now obsolete kubeconfig-path
in the Kind related docs and make machineries. Kind now
uses contexts and sets it properly, so no need to set
KUBECONFIG now.

Signed-off-by: Nikolay Nikolaev <nnikolay@vmware.com>