Rename config env (#1960)
* Rename ADVERTISE_NSE_NAME to ENDPOINT_NETWORK_SERVICE

Part of the effort to unify the environment variables naming.

Signed-off-by: Nikolay Nikolaev <nnikolay@vmware.com>

* Rename ADVERTISE_NSE_LABELS to ENDPOINT_LABELS

Part of the effort to unify the environment variables naming.

Signed-off-by: Nikolay Nikolaev <nnikolay@vmware.com>

* Rename OUTGOING_NSC_NAME to CLIENT_NETWORK_SERVICE

Part of the effort to unify the environment variables naming.

Signed-off-by: Nikolay Nikolaev <nnikolay@vmware.com>

* Rename OUTGOING_NSC_LABELS to CLIENT_LABELS

Part of the effort to unify the environment variables naming.

Signed-off-by: Nikolay Nikolaev <nnikolay@vmware.com>

* SDK: unify internal env constants naming

nscInterfaceNameEnv and namespaceEnv. The latter is also fixing
a linter issue, where an exported name shoudl have a comment. And
we actually don't need to export that constant.

Signed-off-by: Nikolay Nikolaev <nnikolay@vmware.com>