Rough draft for changes for path+healing+forwarder changes.
Note: This whole PR is to large, and will need to be pulled
into smaller pieces.

The purpose of this PR is to get directional feedback.

Please also note this is not *quite* finished yet.  The goal
is to get feedback, and start breaking pieces of this off into
things which can be merged in easy stages and other folks
can start working on.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>