Fix: Sometimes VPN tests not passing because ping is not a success with 1 attempt (#2021)
* fix vpn tests via ethernet context

Sometimes vpn test not passing because ping is not a success with 1 attempt. By logs from https://circleci.com/gh/networkservicemesh/networkservicemesh/118345 I found that ARP entry was not set on NSC interface. Because at now EthernetContext using only for remote cases (Before we did not face with ARP problem for local cases).

Signed-off-by: denis-tingajkin <denis.tingajkin@xored.com>