Add podman support for local builds and deployments
This change adds support for podman. This is useful in environments
where the container runtime is CRI-O instead of Docker. Support is
achieved by enabling Makefile's and build scripts to test for the
docker CLI and fall back to podman if docker is not available, relying
on CLI argument compatibility between the two tools.

Signed-off-by: Ryan Tidwell <rtidwell@suse.com>