Bump spire-agent and spire-server to 0.11.0
Should fix issue #2183. The issue seems to be caused by incorrect cgroup
namespace path parsing in spire-agent:0.9.0.

The cgroup namespace path of a container depends on its QoS class.
Consider the following cgroup namespaces examples:

/docker/0ae0b671d2553a17173876dfed1986eb3a3fbd9cbaad878ea100db8b32a46280/kubepods/besteffort/pod90d75b24-7573-494d-86be-73a1f6c53cf2/45abc7d03395c056a6a79643211272a3e4e28cef9f70e6e4d36283398ac2a101
(of a container with BestEffort QoS class)

/docker/0ae0b671d2553a17173876dfed1986eb3a3fbd9cbaad878ea100db8b32a46280/kubepods/pod02f1e446-526c-46fd-b6be-ae78a807a4c1/96057b397de140bf4a2fd9a6b2bd97a78d0fcd05987f33a36ac0a14e49dbde89
(of a container with a Guranteed QoS class)

Spire-agent 0.9.0 expects at least 4 slashes in the path starting from
`/kubepods/`, which is not the case for the second example.

The parsing bug is fixed in spire-agent version 0.10.1 in
https://github.com/spiffe/spire/pull/1547 .

Since we are bumping the spire version anyway, bump it to the latest
version 0.11.0.

Signed-off-by: Albert Safin <albert.safin@xored.com>