Fix #2245 Create NSM_NAMESPACE if missing on helm-install-nsm (#2246)
* Fix #2245 Create NSM_NAMESPACE if missing on helm install

This change fixes issue #2245 . It updates the helm-nsm-install script to create the NSM_NAMESPACE if it does not already exist. Helm no longer will create namespaces if they don't exist, and so this is required so that the install that happens later in the script does not fail.

Signed-off-by: Chris Johnston <chrijoh@microsoft.com>

* lint: wrap $NSM_NAMESPACE in double quotes to prevent globbing

This change fixes a linting issue which was failing the build. The
NSM_NAMESPACE parameter was not wrapped in double quotes.

See: https://www.shellcheck.net/wiki/SC2086

Signed-off-by: Chris Johnston <chrijoh@microsoft.com>